Fix h100 red and other warnings.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30384 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/buflib.c
@@ -765,7 +765,7 @@ void buflib_print_block_at(struct buflib_context *ctx, int block_num,
         block_num -= 1;
     }
     snprintf(buf, bufsize, "%8p: val: %4ld (%s)",
-                            this, this->val,
+                            this, (long)this->val,
                             this->val > 0? this[3].name:"<unallocated>");
 }
